Revert "Merge branch 'visualization-refactoring'"
This reverts commit 9bb591e84bd1738acdd5db75f3a29896f1714f38,
reversing changes made to 2d62c735973a110b43c35f2901d496f3bce5becf.
