+++ dev/features/hu.elte.txtuml.feature/feature.xml
@@ -262,12 +262,5 @@ This Agreement is governed by the laws of the State of New York and the intellec
          version="0.0.0"
          unpack="false"/>
 
-   <plugin
-         id="hu.elte.txtuml.wizards"
-         download-size="0"
-         install-size="0"
-         version="0.0.0"
-         unpack="false"/>
-
 </feature>
