+++ dev/plugins/hu.elte.txtuml.export.cpp/src/hu/elte/txtuml/export/cpp/wizardz/TxtUMLToCppWizard.java
@@ -18,7 +18,7 @@ import hu.elte.txtuml.export.cpp.EnvironmentNotFoundException;
 import hu.elte.txtuml.export.cpp.Uml2ToCppExporter;
 import hu.elte.txtuml.utils.Pair;
 import hu.elte.txtuml.utils.eclipse.Dialogs;
-import hu.elte.txtuml.utils.eclipse.EditorUtils;
+import hu.elte.txtuml.utils.eclipse.SaveUtils;
 import hu.elte.txtuml.utils.eclipse.WizardUtils;
 
 public class TxtUMLToCppWizard extends Wizard {
@@ -69,7 +69,7 @@ public class TxtUMLToCppWizard extends Wizard {
 			String txtUMLModel = model.getFirst();
 			String txtUMLProject = model.getSecond();
 
-			boolean saveSucceeded = EditorUtils.saveAffectedFiles(getShell(), txtUMLProject, txtUMLModel,
+			boolean saveSucceeded = SaveUtils.saveAffectedFiles(getShell(), txtUMLProject, txtUMLModel,
 					threadManagementDescription.getFullyQualifiedName());
 			if (!saveSucceeded)
 				return false;
