+++ dev/plugins/hu.elte.txtuml.export.javascript/META-INF/MANIFEST.MF
@@ -8,13 +8,16 @@ Bundle-Vendor: ELTE
 Require-Bundle: 
  hu.elte.txtuml.layout.visualizer,
  hu.elte.txtuml.layout.export,
+ hu.elte.txtuml.api.model,
  hu.elte.txtuml.api.layout,
  hu.elte.txtuml.utils,
+ hu.elte.txtuml.utils.eclipse,
  hu.elte.txtuml.export.uml2,
  org.eclipse.debug.ui,
  org.eclipse.jdt.core,
  org.eclipse.jdt.launching,
  org.eclipse.ui.ide,
+ org.eclipse.uml2.uml,
  org.eclipse.emf.ecore,
  org.eclipse.jdt.ui,
  com.google.guava,
@@ -24,6 +27,6 @@ Require-Bundle:
  org.eclipse.core.runtime,
  org.eclipse.persistence.moxy,
  org.eclipse.papyrus.uml.diagram.common,
- hu.elte.txtuml.wizards;bundle-version="0.6.0"
+ hu.elte.txtuml.export.papyrus
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
