+++ dev/plugins/hu.elte.txtuml.export.plantuml/src/hu/elte/txtuml/export/plantuml/handlers/PlantUMLVisualizeSelectedDiagramsHandler.java
@@ -18,7 +18,7 @@ import org.eclipse.jface.wizard.WizardDialog;
 import org.eclipse.ui.handlers.HandlerUtil;
 
 import hu.elte.txtuml.export.plantuml.wizards.PlantUMLVisualizeWizard;
-import hu.elte.txtuml.wizards.VisualizeTxtUMLPage;
+import hu.elte.txtuml.utils.eclipse.wizards.VisualizeTxtUMLPage;
 
 public class PlantUMLVisualizeSelectedDiagramsHandler extends AbstractHandler {
