+++ dev/plugins/hu.elte.txtuml.layout.export/META-INF/MANIFEST.MF
@@ -7,10 +7,10 @@ Bundle-Vendor: ELTE
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: hu.elte.txtuml.layout.export
 Require-Bundle: hu.elte.txtuml.utils,
+ hu.elte.txtuml.layout.visualizer,
  hu.elte.txtuml.api.layout,
  hu.elte.txtuml.api.model,
  hu.elte.txtuml.export.uml2,
  hu.elte.txtuml.utils.eclipse,
  org.eclipse.jdt.core,
- org.eclipse.core.runtime,
- hu.elte.txtuml.layout.visualizer
+ org.eclipse.core.runtime
