+++ dev/plugins/hu.elte.txtuml.utils.eclipse/src/hu/elte/txtuml/utils/eclipse/WizardUtils.java
@@ -202,6 +202,10 @@ public class WizardUtils {
 						.filter(mvp -> mvp.getValueKind() == IMemberValuePair.K_CLASS)
 						.flatMap(mvp -> Stream.of(mvp.getValue())).collect(Collectors.toList());
 
+				if (annotValues.isEmpty()) {
+					throw new NoSuchElementException("Group is empty.");
+				}
+
 				for (Object val : annotValues) {
 					List<Object> annotations = new ArrayList<>();
 					if (val instanceof String) {
