--- dev/plugins/hu.elte.txtuml.wizards/.gitignore
@@ -1,3 +1,0 @@
-bin/
-target/
-.settings/
