--- dev/plugins/hu.elte.txtuml.wizards/META-INF/MANIFEST.MF
@@ -1,21 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: txtUML exportation wizards for Eclipse
-Bundle-SymbolicName: hu.elte.txtuml.wizards;singleton:=true
-Bundle-Version: 0.6.0.qualifier
-Bundle-Vendor: ELTE
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Export-Package: hu.elte.txtuml.wizards
-Require-Bundle: org.eclipse.jdt.core,
- org.eclipse.core.runtime,
- hu.elte.txtuml.utils,
- org.eclipse.core.resources,
- org.eclipse.jface,
- org.eclipse.ui.ide,
- org.eclipse.ui.workbench,
- hu.elte.txtuml.export.uml2,
- hu.elte.txtuml.layout.export,
- hu.elte.txtuml.utils.eclipse;bundle-version="0.6.0",
- org.eclipse.uml2.uml;bundle-version="5.2.2"
-Bundle-ActivationPolicy: lazy
-Bundle-Activator: hu.elte.txtuml.wizards.Activator
